kEqz and kLtz gates are added

+++ src/dirigent/types.h
@@ -11,8 +11,8 @@ using Ring = uint64_t;
 constexpr uint64_t FRACTION = 16;
 
 // using Field = NTL::ZZ_p;
-// using Field = uint64_t;
-using Field = int;
+using Field = uint64_t;
+// using Field = int;
 
 class BoolRing {
   bool val_;
